Update docs formatting
For better compatibility with https://github.com/funbox/blueprinter

It seems that Apiary itself was VERY PERMISSIVE with the formatting
here, because it seems like most of out doc was non-compliant with the
API Blueprint spec https://apiblueprint.org/documentation/specification.html

So most of these changes are just fixing that stuff.

We also remove the github markdown tables because they seem to be not
supported by blueprinter.

Longer term we need to split out all the copy and explanations from this
doc. The format is really only designed for simple request / response
payloads, not all this fancy formatted copy. It's not a good experience
for the user or the writer.