Quick formatting pass for old content
Blueprinter text rendering is EXTREMELY WACK. The formatting of the
markdown DOES impact visual rendering of the text on the page (which is
contrary to normal markdown). For best visual appearance we'd want to
have no line breaks in the content and let the HTML wrap itself, but
this is terrible to manually edit because of the long lines.

So here I'm hard wrapping lines to 120 lines which approximately matches
the rendered width of the lines Blueprinter makes. It's a bit of an
unhappy middle ground, but at least it's consistently terrible now