Expose placement data at job level
For some unknown reason the existing Factory Job API has a
'personalisation' key which shows the ROSTER information [1]. I don't
know why it's called this and I don't know what's using it (Embed
doesn't).

This new 'personalisation' key will be the CUSTOMISED PLACEMENTS data we
have for each job. I'm proposing to not call it 'placements' since not
all placements will have a customisation, which might be confusing.
'customised_placements' might be best.

For each customised placement in this object we have all the same
metadata that we show on the job ticket, as well as the artwork ID. I'm
then removing the stupid separate 'artwork_ids' and 'artwork_dimensions'
objects which are redundant.

[1] https://github.com/unmadeworks/factory/blob/cc75106e6cfb7461b6f0ed44c4b7b2eac8c3b44d/unmade_factory/job/serializers.py#L286