Remove design source ID from Job data
We currently only have the editor design on Factory, which isn't really
the design the user would want to check on Embed. So we're deciding to
remove the reference to the design entirely on the Factory side since it
was only a nice to have really. In future we can clarify this, maybe by
linking up to the order item insted.